requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
backcall==0.2.0           # via ipython
certifi==2020.6.20        # via sentry-sdk
decorator==4.4.2          # via ipython, traitlets
django-autocomplete-light==3.5.1  # via -r requirements.in
django-crispy-forms==1.7.2  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==3.0.3  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-imagefield==0.12.4  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-tables2==2.3.1     # via -r requirements.in
django-widget-tweaks==1.4.8  # via -r requirements.in
django==2.2.10            # via -r requirements.in, django-filter, django-model-utils, django-tables2
et-xmlfile==1.0.1         # via openpyxl
factory-boy==2.12.0       # via -r requirements.in
faker==4.1.2              # via factory-boy
gunicorn==20.0.4          # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements.in
jdcal==1.4.1              # via openpyxl
jedi==0.17.2              # via ipython
openpyxl==3.0.5           # via tablib
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via -r requirements.in, django-imagefield
prompt-toolkit==3.0.6     # via ipython
psycopg2-binary==2.8.5    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pygments==2.6.1           # via ipython
python-dateutil==2.8.1    # via faker
pytz==2019.1              # via django
sentry-sdk==0.16.5        # via -r requirements.in
six==1.15.0               # via django-autocomplete-light, django-extensions, python-dateutil, traitlets
sqlparse==0.3.0           # via django
tablib[xlsx]==2.0.0       # via -r requirements.in
text-unidecode==1.3       # via faker
traitlets==4.3.3          # via ipython
urllib3==1.25.10          # via sentry-sdk
wcwidth==0.2.5            # via prompt-toolkit
whitenoise==5.1.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
